+++ tests/acceptance/01checkoutWithPaymorrowTest.php
@@ -30,41 +30,18 @@
  * the OXID module for Paymorrow payment. If not, see <http://www.gnu.org/licenses/>.
  */
 
-if (!defined('PAYMORROW_CONFIG')) {
-    include(dirname(__FILE__) . DIRECTORY_SEPARATOR . '..' . DIRECTORY_SEPARATOR . 'additional.inc.php');
-}
+include_once 'PaymorrowAcceptanceTestCase.php';
 
 /**
  * Class Acceptance_01checkoutWithPaymorrowTest.
  * Test different checkout cases using Paymorrow payment methods.
  */
-class Acceptance_01checkoutWithPaymorrowTest extends OxidEsales\TestingLibrary\AcceptanceTestCase
+class Acceptance_01checkoutWithPaymorrowTest extends PaymorrowAcceptanceTestCase
 {
-
-    /**
-     * @var bool
-     */
-    protected $_blStartMinkSession = false;
-
-    /**
-     * @var null|\Behat\Mink\Session
-     */
-    protected $_oMinkSession = null;
-
-    /**
-     * @var null|DocumentElement
-     */
-    protected $_oPage = null;
-
     protected $_sShopIdParam = '';
 
     /** @var string Language id. */
     protected $translateLanguageId = '0';
-
-    /**
-     * How much more time wait for these tests.
-     */
-    protected $_iWaitTimeMultiplier = 9;
 
     /**
      * Prepare shop for testing.
@@ -73,18 +50,6 @@ class Acceptance_01checkoutWithPaymorrowTest extends OxidEsales\TestingLibrary\A
     {
         parent::setUp();
 
-        if ( !PAYMORROW_CONFIG ) {
-            exit( 'Paymorrow tests are not configured. Please set it up in "paymorrow_config.php" file.' . PHP_EOL );
-        }
-
-        $this->activateTheme('azure');
-
-        // Start Mink session wuth Selenium driver
-        $this->startMinkSession( 'selenium' );
-
-        // Get page
-        $this->_oMinkSession = $this->getMinkSession();
-        $this->_oPage        = $this->_oMinkSession->getPage();
         $this->getTranslator()->setLanguage(0);
 
         if (isSUBSHOP && oxSHOPID) {
@@ -95,7 +60,6 @@ class Acceptance_01checkoutWithPaymorrowTest extends OxidEsales\TestingLibrary\A
         // Open shop
         $this->openShop();
     }
-
 
     /*---------------------------------------*
      *- Test with different payment methods -*